+++ example/lib/conditional.dart
@@ -30,7 +30,7 @@ class App extends StatelessWidget {
     GoRoute(
       pattern: '/',
       builder: (context, state) => MaterialPage<HomePage>(
-        key: const ValueKey('HomePage'),
+        key: state.pageKey,
         child: HomePage(families: Families.data),
       ),
       routes: [
@@ -40,7 +40,7 @@ class App extends StatelessWidget {
             final family = Families.family(state.params['fid']!);
 
             return MaterialPage<FamilyPage>(
-              key: ValueKey(family),
+              key: state.pageKey,
               child: FamilyPage(family: family),
             );
           },
@@ -52,7 +52,7 @@ class App extends StatelessWidget {
                 final person = family.person(state.params['pid']!);
 
                 return MaterialPage<PersonPage>(
-                  key: ValueKey(person),
+                  key: state.pageKey,
                   child: PersonPage(family: family, person: person),
                 );
               },
@@ -67,9 +67,9 @@ class App extends StatelessWidget {
   final _loggedOutRoutes = [
     GoRoute(
       pattern: '/',
-      builder: (context, state) => const MaterialPage<LoginPage>(
-        key: ValueKey('LoginPage'),
-        child: LoginPage(),
+      builder: (context, state) => MaterialPage<LoginPage>(
+        key: state.pageKey,
+        child: const LoginPage(),
       ),
     ),
   ];
@@ -80,7 +80,7 @@ class App extends StatelessWidget {
 
   Page<dynamic> _errorBuilder(BuildContext context, GoRouterState state) =>
       MaterialPage<ErrorPage>(
-        key: const ValueKey('ErrorPage'),
+        key: state.pageKey,
         child: ErrorPage(message: state.error.toString()),
       );
 }