+++ example/lib/query_params.dart
@@ -34,7 +34,7 @@ class App extends StatelessWidget {
         GoRoute(
           pattern: '/',
           builder: (context, state) => MaterialPage<HomePage>(
-            key: const ValueKey('HomePage'),
+            key: state.pageKey,
             child: HomePage(families: Families.data),
           ),
           routes: [
@@ -43,7 +43,7 @@ class App extends StatelessWidget {
               builder: (context, state) {
                 final family = Families.family(state.params['fid']!);
                 return MaterialPage<FamilyPage>(
-                  key: ValueKey(family),
+                  key: state.pageKey,
                   child: FamilyPage(family: family),
                 );
               },
@@ -54,7 +54,7 @@ class App extends StatelessWidget {
                     final family = Families.family(state.params['fid']!);
                     final person = family.person(state.params['pid']!);
                     return MaterialPage<PersonPage>(
-                      key: ValueKey(person),
+                      key: state.pageKey,
                       child: PersonPage(family: family, person: person),
                     );
                   },
@@ -66,7 +66,7 @@ class App extends StatelessWidget {
         GoRoute(
           pattern: '/login',
           builder: (context, state) => MaterialPage<LoginPage>(
-            key: const ValueKey('LoginPage'),
+            key: state.pageKey,
             // pass the deep link to the LoginPage (if there is one)
             child: LoginPage(from: state.params['from']),
           ),
@@ -75,7 +75,7 @@ class App extends StatelessWidget {
 
   Page<dynamic> _errorBuilder(BuildContext context, GoRouterState state) =>
       MaterialPage<ErrorPage>(
-        key: const ValueKey('ErrorPage'),
+        key: state.pageKey,
         child: ErrorPage(message: state.error.toString()),
       );
 }